+++ modules/Admintoolbox.SecOps/Public/DisMount-ProfileRegistry.ps1
@@ -2,14 +2,11 @@
     .DESCRIPTION
     Loads any local users registry hive into the registry as a friendly name. Useful for modifying a users registry without the need of them being logged in locally.
 
-    .PARAMETER SamAccountName
-    SamAccountName of the domain account that you are mounting the local registry for.
-
-    .PARAMETER DomainController
-    Domain Controller that has Powershell Remoting enabled on it. For importing a session with the Active directory Module.
+    .PARAMETER userprofile
+    Specify the previouslymounted userprofile to unmount.
 
     .EXAMPLE
-    DisMount-ProfileRegistry -SamAccountName JohnS
+    DisMount-ProfileRegistry -userprofile johns
 
     DisMounts the Mounted ProfileRegistry
 
@@ -21,9 +18,9 @@ function DisMount-ProfileRegistry {
 
     [CmdletBinding()]
     Param (
-        [Parameter(Mandatory = $true)]$SamAccountName
+        [Parameter(Mandatory = $true)]$userprofile
     )
 
     #DisMount registry key for user
-    reg.exe unload HKU\$SamAccountName
+    reg.exe unload HKU\$userprofile
 }