+++ modules/Admintoolbox.SecOps/Changelog.md
@@ -17,3 +17,4 @@
 * **1.4.0.2** CI Maintenance Release
 * **1.4.0.3** CI Maintenance Release
 * **1.4.0.4** CI Maintenance Release
+* **1.5.0.0** Added functions for loading profile registries.